# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Estimator Platform API
  version: 1.0.0
  description: Multi-tenant estimator + dashboard SaaS platform API.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
...
paths:
...
  /tenants/signup:
    post:
      tags: [Tenants]
      summary: Create a new tenant and owner user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, ownerEmail]
              properties:
                name:
                  type: string
                slug:
                  type: string
                  pattern: "^[a-z0-9-]+$"
                ownerEmail:
                  type: string
                  format: email
      responses:
        "201":
          description: Tenant and owner created
        "409":
          description: Slug already taken
  /tenants/me:
    get:
      tags: [Tenants]
      summary: Get current tenant details
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Tenant details
  /users:
    get:
      tags: [Users]
      summary: List users in current tenant
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User list
    post:
      tags: [Users]
      summary: Create/invite a user (admin+)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: User created
  /users/{id}:
    delete:
      tags: [Users]
      summary: Delete a user (owner only)
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User deleted
...
  /submissions:
    post:
      tags: [Submissions]
      summary: Submit an estimate (public, called by widget)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmission"
      responses:
        "201":
          description: Submission created with estimate
        "422":
          description: Out of service area
    get:
      tags: [Submissions]
      summary: List submissions (authenticated)
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Submission list
...
  /exports/submissions:
    get:
      tags: [Exports]
      summary: Export submissions as CSV
      security:
        - sessionCookie: []
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ep_session
  schemas:
...
    CreateUser:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
...
    CreateSubmission:
      type: object
      required: [estimatorPublicKey, lead, answers]
      properties:
        estimatorPublicKey:
          type: string
        lead:
          type: object
          required: [email, zip]
          properties:
            email:
              type: string
              format: email
            zip:
              type: string
            name:
              type: string
            phone:
              type: string
        answers:
          type: object
...